JpaCursorConfiguration

+++ src/main/resources/schema-mysql.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `customer`
 (
-    `id`        mediumint(8) unsigned NOT NULL auto_increment,
+    `id` bigint unsigned NOT NULL auto_increment,
     `firstName` varchar(255) default NULL,
     `lastName`  varchar(255) default NULL,
     `birthdate` varchar(255),
